Add support for wstring.
miniutf can now convert from and to wstring. The assumption here is that wstring is encoded as UTF-16. Since it’s either 2 (MSVC) or 4 bytes (Clang, GCC) depending on compiler, we can’t assume it holds UTF-32. If it holds UTF-8, the conversion is trivial and minutf is not needed. In practice every function that took or returned `u16string` now has a `wstring` match.

I used templates to ensure we don’t have repetitive code. For simplicity reasons, no templates are exposed in the header though.

Tests have been added, however the test spec is quite minimal.